requirements.txt
pyzmq
prettytable==0.7.2